terminal/requirements.txt
transformers>=4.40.0
python-dotenv
